moved all '/'route functions to helpers.js
